+++ ChangeLog.md
@@ -9,10 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added 
 
+* `Keelung.Syntax.Monad`
+    * Functions for requesting fresh varibles: `freshVar`, `freshVarField`, `freshVarBool`, and `freshVarUInt`.
+
 ### Fixed
 
 ### Removed
 
-* Function `uintToBool`
-* Function `fieldToBool`
+* `Keelung.Syntax`
+    * Function `uintToBool`
+    * Function `fieldToBool`